+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.3.0] - 2023-05-25
+### Added
+- New option `--create`, to create unreleased versions [#31].
+
+### Fixed
+- Allow empty values for `--release` option
+
 ## [2.2.1] - 2023-01-25
 ### Fixed
 - Allow list elements in the descriptions of releases and changelogs [#30].
@@ -44,7 +51,9 @@ New version merging Deno and Node code using Deno's `dnt` package.
 [#27]: https://github.com/oscarotero/keep-a-changelog/issues/27
 [#28]: https://github.com/oscarotero/keep-a-changelog/issues/28
 [#30]: https://github.com/oscarotero/keep-a-changelog/issues/30
+[#31]: https://github.com/oscarotero/keep-a-changelog/issues/31
 
+[2.3.0]: https://github.com/oscarotero/keep-a-changelog/compare/v2.2.1...v2.3.0
 [2.2.1]: https://github.com/oscarotero/keep-a-changelog/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/oscarotero/keep-a-changelog/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/oscarotero/keep-a-changelog/compare/v2.0.1...v2.1.0